seperate animation file: in progress of fixing issue
